refactor: move custom AE implementation from hooks to extended class
